+++ packages/devtools/src/utils.ts
@@ -3,10 +3,10 @@ import logger from '@wdio/logger'
 import { commandCallStructure, isValidParameter, getArgumentType, canAccess } from '@wdio/utils'
 import { WebDriverProtocol, CommandParameters, CommandPathVariables, ElementReference } from '@wdio/protocols'
 import type { Logger } from '@wdio/logger'
-import type { ElementHandle } from 'puppeteer-core/lib/cjs/puppeteer/common/JSHandle'
-import type { Browser } from 'puppeteer-core/lib/cjs/puppeteer/common/Browser'
-import type { Frame } from 'puppeteer-core/lib/cjs/puppeteer/common/FrameManager'
-import type { Page } from 'puppeteer-core/lib/cjs/puppeteer/common/Page'
+import type { ElementHandle } from 'puppeteer-core/lib/cjs/puppeteer/common/ElementHandle'
+import type { Browser } from 'puppeteer-core/lib/cjs/puppeteer/api/Browser'
+import type { Frame } from 'puppeteer-core/lib/cjs/puppeteer/common/Frame'
+import type { Page } from 'puppeteer-core/lib/cjs/puppeteer/api/Page'
 
 import cleanUp from './scripts/cleanUpSerializationSelector'
 import { ELEMENT_KEY, SERIALIZE_PROPERTY, SERIALIZE_FLAG, ERROR_MESSAGES, PPTR_LOG_PREFIX } from './constants'
@@ -110,10 +110,10 @@ export async function findElement (
         await waitForFn.call(context, value, { timeout: implicitTimeout })
     }
 
-    let element
+    let element: ElementHandle<Element> | null = null
     try {
         element = using === 'xpath'
-            ? (await context.$x(value))[0]
+            ? (await context.$x(value))[0] as ElementHandle<Element>
             : await context.$(value)
     } catch (err: any) {
         /**
@@ -151,7 +151,7 @@ export async function findElements (
     }
 
     const elements = using === 'xpath'
-        ? await context.$x(value)
+        ? await context.$x(value) as ElementHandle<Element>[]
         : await context.$$(value)
 
     if (elements.length === 0) {