+++ packages/wdio-devtools-service/tests/__mocks__/ws.ts
@@ -1,0 +1,7 @@
+export default class WebSocketMock {
+    on = jest.fn().mockReturnValue({ off: jest.fn() })
+    off = jest.fn()
+    send = jest.fn()
+
+    constructor (public wsUrl: string) {}
+}